Expose transport message interfaces so that channel developers can develop richer channels
