+++ src/index.ts
@@ -1,3 +1,4 @@
 export { slot, Slot } from './Slot'
 export { EventDeclaration, combineEvents, createEventBus } from './Events'
 export { Channel, GenericChannel } from './Channel'
+export { TransportMessage } from './Message'
